Resolve issue with missing Control Info on new window launch
When opening the application in new windows, the 'Application.Current.OpenWindow()' method lacks support for passing navigation arguments (navArgs). As a workaround, implemented manual provision of control information to the page and enforced reload with the 'Loaded' event on each affected page. This ensures consistent behavior and preserves data integrity across window launches.